Why is it 0

+++ src/ArchimedesPiMethod.java
@@ -1,9 +1,9 @@
 import java.lang.Math;
 public class ArchimedesPiMethod {
     public static void main(String[] args) {
-        int n = 8;
-        int b = 360 / n;
-        int a = (1/2) * b;
+        float n = 8;
+        float b = 360 / n;
+        float a = (1/2f) * b;
         double newA = Math.toRadians(a);
         double s = 2 * Math.sin(newA);
         double p = n * s;
